+++ plone/tiles/esi.py
@@ -1,14 +1,24 @@
 # -*- coding: utf-8 -*-
-
 from plone.tiles.interfaces import ESI_HEADER
 from plone.tiles.interfaces import ESI_HEADER_KEY
 from plone.tiles.interfaces import IESIRendered
+from plone.tiles.interfaces import ITileType
 from plone.tiles.tile import PersistentTile
 from plone.tiles.tile import Tile
+from Products.Five import BrowserView
+from zExceptions import Unauthorized
+from zope.component import queryUtility
 from zope.interface import implementer
+try:
+    from AccessControl.security import checkPermission
+except ImportError:
+    from zope.security import checkPermission
 
+import os
 import re
+import transaction
 
+X_FRAME_OPTIONS = os.environ.get('PLONE_X_FRAME_OPTIONS', 'SAMEORIGIN')
 
 HEAD_CHILDREN = re.compile(r'<head[^>]*>(.*)</head>', re.I | re.S)
 BODY_CHILDREN = re.compile(r'<body[^>]*>(.*)</body>', re.I | re.S)
@@ -60,7 +70,9 @@ class ConditionalESIRendering(object):
             if self.head:
                 mode = 'esi-head'
             return ESI_TEMPLATE.format(
-                url=self.request.getURL(),
+                url=(self.request.get('PATH_INFO') and
+                     self.request.get('PATH_INFO').replace(' ', '%20') or
+                     self.request.getURL()),
                 queryString=self.request.get('QUERY_STRING', ''),
                 esiMode=mode
             )
@@ -96,22 +108,30 @@ class ESIPersistentTile(ConditionalESIRendering, PersistentTile):
 
 # ESI views
 
-class ESIHead(object):
+class ESIHead(BrowserView):
     """Render the head portion of a tile independently.
     """
-
-    def __init__(self, context, request):
-        self.tile = context
-        self.request = request
 
     def __call__(self):
         """Return the children of the <head> tag as a fragment.
         """
+        # Check for the registered view permission
+        try:
+            type_ = queryUtility(ITileType, self.context.__name__)
+            permission = type_.view_permission
+        except AttributeError:
+            permission = None
+        if permission:
+            if not checkPermission(permission, self.context):
+                raise Unauthorized()
 
         if self.request.getHeader(ESI_HEADER):
             del self.request.environ[ESI_HEADER_KEY]
 
-        document = self.tile()  # render the tile
+        document = self.context()  # render the tile
+
+        # Disable the theme so we don't <html/>-wrapped
+        self.request.response.setHeader('X-Theme-Disabled', '1')
 
         match = HEAD_CHILDREN.search(document)
         if not match:
@@ -119,24 +139,70 @@ class ESIHead(object):
         return match.group(1).strip()
 
 
-class ESIBody(object):
+class ESIBody(BrowserView):
     """Render the head portion of a tile independently.
     """
-
-    def __init__(self, context, request):
-        self.tile = context
-        self.request = request
 
     def __call__(self):
         """Return the children of the <head> tag as a fragment.
         """
+        # Check for the registered view permission
+        try:
+            type_ = queryUtility(ITileType, self.context.__name__)
+            permission = type_.view_permission
+        except AttributeError:
+            permission = None
+        if permission:
+            if not checkPermission(permission, self.context):
+                raise Unauthorized()
 
         if self.request.getHeader(ESI_HEADER):
             del self.request.environ[ESI_HEADER_KEY]
 
-        document = self.tile()  # render the tile
+        document = self.context()  # render the tile
+
+        # Disable the theme so we don't <html/>-wrapped
+        self.request.response.setHeader('X-Theme-Disabled', '1')
 
         match = BODY_CHILDREN.search(document)
         if not match:
             return document
         return match.group(1).strip()
+
+
+class ESIProtectTransform(object):
+    """Replacement transform for plone.protect's ProtectTransform,
+    because ESI tile responses' HTML should not be transformed to
+    avoid wrapping them with <html>-tag
+    """
+
+    order = 9000
+
+    def __init__(self, published, request):
+        self.published = published
+        self.request = request
+
+    def transform(self, result, encoding):
+        from plone.protect.interfaces import IDisableCSRFProtection
+        # clickjacking protection from plone.protect
+        if X_FRAME_OPTIONS:
+            if not self.request.response.getHeader('X-Frame-Options'):
+                self.request.response.setHeader(
+                    'X-Frame-Options', X_FRAME_OPTIONS)
+        # drop X-Tile-Url
+        if 'x-tile-url' in self.request.response.headers:
+            del self.request.response.headers['x-tile-url']
+        # ESI requests are always GET request and should not mutate DB
+        # unless they provide IDisableCSRFProtection
+        if not IDisableCSRFProtection.providedBy(self.request):
+            transaction.abort()
+        return None
+
+    def transformBytes(self, result, encoding):
+        return self.transform(result, encoding)
+
+    def transformUnicode(self, result, encoding):
+        return self.transform(result, encoding)
+
+    def transformIterable(self, result, encoding):
+        return self.transform(result, encoding)